requirements.txt
alembic==1.4.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
databricks-cli==0.15.0
docker==5.0.3
entrypoints==0.3
filelock==3.3.0
Flask==2.0.2
gitdb==4.0.7
GitPython==3.1.24
greenlet==1.1.2
grpcio==1.41.0
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
Mako==1.1.5
MarkupSafe==2.0.1
mlflow==1.20.2
msgpack==1.0.2
numpy==1.21.2
packaging==21.0
pandas==1.3.3
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.3
protobuf==3.18.1
psycopg2==2.9.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.3
pywin32==227
PyYAML==5.4.1
querystring-parser==1.2.4
ray==1.7.0
redis==3.5.3
requests==2.26.0
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
smmap==4.0.0
SQLAlchemy==1.4.25
sqlparse==0.4.2
tabulate==0.8.9
tensorboardX==2.4
threadpoolctl==3.0.0
tune-sklearn==0.4.1
typing-extensions==3.10.0.2
urllib3==1.26.7
waitress==2.0.0
websocket-client==1.2.1
Werkzeug==2.0.2
xgboost==1.4.2
zipp==3.6.0
